+++ setup_sklearnex.py
@@ -85,7 +85,6 @@ with open("README.md", "r", encoding="utf8") as f:
 packages_with_tests = [
     "sklearnex",
     "sklearnex.cluster",
-    "sklearnex.covariance",
     "sklearnex.decomposition",
     "sklearnex.ensemble",
     "sklearnex.glob",
@@ -95,6 +94,7 @@ packages_with_tests = [
     "sklearnex.model_selection",
     "sklearnex.neighbors",
     "sklearnex.preview",
+    "sklearnex.preview.covariance",
     "sklearnex.preview.cluster",
     "sklearnex.preview.decomposition",
     "sklearnex.preview.linear_model",
