+++ sklearnex/__init__.py
@@ -32,7 +32,6 @@ __all__ = [
     "basic_statistics",
     "cluster",
     "config_context",
-    "covariance",
     "decomposition",
     "ensemble",
     "get_config",
