+++ sklearnex/preview/__init__.py
@@ -15,4 +15,4 @@
 # limitations under the License.
 # ==============================================================================
 
-__all__ = ["cluster", "decomposition", "linear_model"]
+__all__ = ["cluster", "covariance", "decomposition", "linear_model"]
